backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
python-multipart==0.0.6
openai>=1.0.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4